Migrate InternetConnectivity check to new API
