+++ apps/settings/lib/AppInfo/Application.php
@@ -50,6 +50,7 @@ use OCA\Settings\Search\SectionSearch;
 use OCA\Settings\Search\UserSearch;
 use OCA\Settings\SetupChecks\CheckUserCertificates;
 use OCA\Settings\SetupChecks\DefaultPhoneRegionSet;
+use OCA\Settings\SetupChecks\InternetConnectivity;
 use OCA\Settings\SetupChecks\LegacySSEKeyFormat;
 use OCA\Settings\SetupChecks\PhpDefaultCharset;
 use OCA\Settings\SetupChecks\PhpOutdated;
@@ -146,13 +147,14 @@ class Application extends App implements IBootstrap {
 			);
 		});
 		$context->registerSetupCheck(CheckUserCertificates::class);
+		$context->registerSetupCheck(DefaultPhoneRegionSet::class);
+		$context->registerSetupCheck(InternetConnectivity::class);
 		$context->registerSetupCheck(LegacySSEKeyFormat::class);
 		$context->registerSetupCheck(PhpDefaultCharset::class);
 		$context->registerSetupCheck(PhpOutdated::class);
 		$context->registerSetupCheck(PhpOutputBuffering::class);
-		$context->registerSetupCheck(SupportedDatabase::class);
-		$context->registerSetupCheck(DefaultPhoneRegionSet::class);
 		$context->registerSetupCheck(ReadOnlyConfig::class);
+		$context->registerSetupCheck(SupportedDatabase::class);
 
 		$context->registerUserMigrator(AccountMigrator::class);
 	}
